// packages/apis/https/enums/httpEnum.ts
/**
 * @description:  http常量
 * @author: gbeata
 */
export enum ResultEnum {
  SUCCESS = 0,
  ERROR = -1,
  TIMEOUT = 401,
  TYPE = 'success',
}
/**
 * @description:  请求方式
 * @author: gbeata
 */
export enum RequestEnum {
  GET = 'GET',
  POST = 'POST',
  PUT = 'PUT',
  DELETE = 'DELETE',
}

/**
 * @description:  常用Content Type
 * @author: gbeata
 */
export enum ContentTypeEnum {
  // json
  JSON = 'application/json;charset=UTF-8',
  // form-data qs
  FORM_URLENCODED = 'application/x-www-form-urlencoded;charset=UTF-8',
  // form-data  upload
  FORM_DATA = 'multipart/form-data;charset=UTF-8',
  // stream data upload
  STREAM_DATA = 'application/octet-stream',
}

export enum StorageEnum {
  Token = 'token',
}